+++ browser/brave_tab_helpers.cc
@@ -18,7 +18,6 @@
 #include "brave/browser/misc_metrics/page_metrics_tab_helper.h"
 #include "brave/browser/misc_metrics/process_misc_metrics.h"
 #include "brave/browser/ntp_background/ntp_tab_helper.h"
-#include "brave/browser/onboarding/onboarding_tab_helper.h"
 #include "brave/browser/skus/skus_service_factory.h"
 #include "brave/browser/ui/bookmark/brave_bookmark_tab_helper.h"
 #include "brave/components/ai_chat/core/common/buildflags/buildflags.h"
@@ -49,6 +48,7 @@
 #endif
 
 #if !BUILDFLAG(IS_ANDROID)
+#include "brave/browser/onboarding/onboarding_tab_helper.h"
 #include "brave/browser/ui/brave_shields_data_controller.h"
 #include "chrome/browser/ui/thumbnails/thumbnail_tab_helper.h"
 #endif
@@ -117,7 +117,7 @@ void AttachTabHelpers(content::WebContents* web_contents) {
   // Add tab helpers here unless they are intended for android too
   BraveBookmarkTabHelper::CreateForWebContents(web_contents);
   brave_shields::BraveShieldsDataController::CreateForWebContents(web_contents);
-  onboarding::OnboardingTabHelper::CreateForWebContents(web_contents);
+  OnboardingTabHelper::MaybeCreateForWebContents(web_contents);
   ThumbnailTabHelper::CreateForWebContents(web_contents);
 #endif
