+++ browser/ui/BUILD.gn
@@ -829,10 +829,10 @@ source_set("ui") {
       "views/brave_actions/brave_rewards_action_view.h",
       "views/brave_actions/brave_shields_action_view.cc",
       "views/brave_actions/brave_shields_action_view.h",
-      "views/brave_help_bubble/brave_help_bubble_delegate.cc",
-      "views/brave_help_bubble/brave_help_bubble_delegate.h",
-      "views/brave_help_bubble/brave_help_bubble_view.cc",
-      "views/brave_help_bubble/brave_help_bubble_view.h",
+      "views/brave_help_bubble/brave_help_bubble_delegate_view.cc",
+      "views/brave_help_bubble/brave_help_bubble_delegate_view.h",
+      "views/brave_help_bubble/brave_help_bubble_host_view.cc",
+      "views/brave_help_bubble/brave_help_bubble_host_view.h",
       "views/frame/vertical_tab_strip_region_view.cc",
       "views/frame/vertical_tab_strip_region_view.h",
       "views/frame/vertical_tab_strip_root_view.cc",
