+++ browser/ui/views/frame/brave_browser_view.h
@@ -53,7 +53,7 @@ class SidebarContainerView;
 class WalletButton;
 class ViewShadow;
 class VerticalTabStripWidgetDelegateView;
-class BraveHelpBubbleView;
+class BraveHelpBubbleHostView;
 
 class BraveBrowserView : public BrowserView,
                          public commands::AcceleratorService::Observer {
@@ -102,7 +102,7 @@ class BraveBrowserView : public BrowserView,
   vertical_tab_strip_widget_delegate_view() {
     return vertical_tab_strip_widget_delegate_view_;
   }
-  void ShowBraveHelpBubbleView(const std::u16string text);
+  void ShowBraveHelpBubbleView(const std::string& text) override;
 
   // commands::AcceleratorService:
   void OnAcceleratorsChanged(const commands::Accelerators& changed) override;
@@ -159,7 +159,7 @@ class BraveBrowserView : public BrowserView,
   void UpdateSideBarHorizontalAlignment();
 
   bool closing_confirm_dialog_activated_ = false;
-  base::WeakPtr<BraveHelpBubbleView> brave_help_bubble_view_;
+  base::WeakPtr<BraveHelpBubbleHostView> brave_help_bubble_view_;
   raw_ptr<SidebarContainerView> sidebar_container_view_ = nullptr;
   raw_ptr<views::View> sidebar_separator_view_ = nullptr;
   raw_ptr<views::View> contents_background_view_ = nullptr;
